move insert to dml
